+++ lib/QECCore/src/codes/quantum/css.jl
@@ -12,11 +12,13 @@ struct CSS <: AbstractCSSCode
     """The parity check matrix of the `Z` stabilizers."""
     Hz::Matrix{Bool}
     function CSS(Hx, Hz)
-        n = size(Hx, 2)
-        if n != size(Hz, 2) error("When constructing a CSS quantum code, the two classical codes are required to have the same block size") end
+        owned_Hx = Matrix{Bool}(Hx)
+        owned_Hz = Matrix{Bool}(Hz)
+        n = size(owned_Hx, 2)
+        if n != size(owned_Hz, 2) error("When constructing a CSS quantum code, the two classical codes are required to have the same block size") end
         #if size(Hx,1)+size(Hz,1) >= n error("When constructing a CSS quantum code, the total number of checks (rows) in the parity checks of the two classical codes have to be lower than the block size (the number of columns).") end
-        check_allrowscommute(Hx, Hz) || error("The CSS code just created is invalid -- its rows do not commute. This is either a bug in this library, or an inconsistent parity check matrices were provided to the CSS constructor.")
-        new(Hx, Hz)
+        check_allrowscommute(owned_Hx, owned_Hz) || error("The CSS code just created is invalid -- its rows do not commute. This is either a bug in this library, or an inconsistent parity check matrices were provided to the CSS constructor.")
+        new(owned_Hx, owned_Hz)
     end
 end
 
@@ -30,18 +32,24 @@ function check_allrowscommute(Hx, Hz)
     return true
 end
 
-function parity_matrix(c::CSS)
-    extended_Hx = Matrix{Bool}(vcat(c.Hx, zeros(size(c.Hz))))
-    extended_Hz = Matrix{Bool}(vcat(zeros(size(c.Hx)), c.Hz))
-    return hcat(extended_Hx, extended_Hz)
+function _css_parity_matrix(Hx::Matrix{Bool}, Hz::Matrix{Bool})
+    return [Hx falses(size(Hx)); falses(size(Hz)) Hz]
 end
 
-parity_matrix_x(c::CSS) = c.Hx
-parity_matrix_z(c::CSS) = c.Hz
+parity_matrix(c::CSS) = _css_parity_matrix(c.Hx, c.Hz)
+
+parity_matrix_xz(c::CSS) = (copy(c.Hx), copy(c.Hz))
+
+parity_matrix_x(c::CSS) = copy(c.Hx)
+parity_matrix_z(c::CSS) = copy(c.Hz)
 
 code_n(c::CSS) = size(c.Hx,2)
 code_s(c::CSS) = size(c.Hx, 1) + size(c.Hz, 1)
 
 
 # Parity matrix for general CSS codes
-parity_matrix(c::AbstractCSSCode) = parity_matrix(CSS(parity_matrix_x(c), parity_matrix_z(c)))
+function parity_matrix(c::AbstractCSSCode)
+    Hx, Hz = parity_matrix_xz(c)
+    css = CSS(Hx, Hz)
+    return _css_parity_matrix(css.Hx, css.Hz)
+end